# doc/api_definition.yaml
swagger: '2.0'
info:
  version: "1.0.0"
  title: DLRN API
host: 127.0.0.1:5000
schemes:
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  basicAuth:
    type: basic
    description: HTTP Basic Authentication. Works over `HTTP` and `HTTPS`
# Describe your paths here
paths:
  /api/last_tested_repo:
    # This is a HTTP operation
    get:
      # Describe this verb here. Note: you can use markdown
      description: |
        Get the last tested repo since a specific time.

        If a ``job_id`` is specified, the order of precedence for the repo returned is:
        - The last tested repo within that timeframe for that CI job.
        - The last tested repo within that timeframe for any CI job, so we can have
          several CIs converge on a single repo.
        - The last "consistent" repo, if no repo has been tested in the timeframe.

        If ``sequential_mode`` is set to true, a different algorithm is used. Another
        parameter ``previous_job_id`` needs to be specified, and the order of
        precedence for the repo returned is:

        - The last tested repo within that timeframe for the CI job described by
        ``previous_job_id``.
        - If no repo for ``previous_job_id`` is found, an error will be returned

        The sequential mode is meant to be used by CI pipelines, where a CI (n) job needs
        to use the same repo tested by CI (n-1).
      # This is array of GET operation parameters:
      parameters:
        - name: params
          in: body
          description: The JSON params to post
          schema:
            type: object
            required:
              - max_age
            properties:
              max_age:
                description: |
                  Maximum age (in hours) for the repo to be considered. Any repo
                  tested or being tested after "now - max_age" will be taken
                  into account. If set to 0, all repos will be considered.
                type: integer
                minimum: 0
              success:
                description: |
                  If set to a value, find repos with a successful/unsuccessful
                  vote (as specified). If not set, any tested repo will be
                  considered.
                type: boolean
              job_id:
                description: |
                  Name of the CI that sent the vote. If not set, no filter will
                  be set on CI.
                type: string
              sequential_mode:
                description: |
                  Use the sequential mode algorithm. In this case, return the
                  last tested repo within that timeframe for the CI job
                  described by previous_job_id. Defaults to false.
                type: boolean
              previous_job_id:
                description: |
                  If sequential_mode is set to true, look for jobs tested by
                  the CI identified by previous_job_id.
                type: string
          required: true
      # Expected responses for this operation:
      responses:
        # Response code
        200:
          description: Successful response
          # A schema describing your response object.
          # Use JSON Schema format
          schema:
            $ref: '#/definitions/Repo'
    post:
      security:
       - basicAuth: []
      # Describe this verb here. Note: you can use markdown
      description: |
        Get the last tested repo since a specific time (optionally for a CI job),
        and add an "in progress" entry in the CI job table for this.

        If a job_id is specified, the order of precedence for the repo returned is:

        - The last tested repo within that timeframe for that CI job.
        - The last tested repo within that timeframe for any CI job, so we can have
          several CIs converge on a single repo.
        - The last "consistent" repo, if no repo has been tested in the timeframe.

        If ``sequential_mode`` is set to true, a different algorithm is used. Another
        parameter ``previous_job_id`` needs to be specified, and the order of
        precedence for the repo returned is:

        - The last tested repo within that timeframe for the CI job described by
          ``previous_job_id``.
        - If no repo for ``previous_job_id`` is found, an error will be returned

        The sequential mode is meant to be used by CI pipelines, where a CI (n) job needs
        to use the same repo tested by CI (n-1).
      # This is array of GET operation parameters:
      parameters:
        - name: params
          in: body
          description: The JSON params to post
          schema:
            type: object
            required:
              - max_age
              - reporting_job_id
            properties:
              max_age:
                description: |
                  Maximum age (in hours) for the repo to be considered. Any repo
                  tested or being tested after "now - max_age" will be taken
                  into account. If set to 0, all repos will be considered.
                type: integer
                minimum: 0
              reporting_job_id:
                description: |
                  Name of the CI that will add the "in progress" entry in the CI
                  job table.
                type: string
              success:
                description: |
                  If set to a value, find repos with a successful/unsuccessful
                  vote (as specified). If not set, any tested repo will be
                  considered.
                type: boolean
              job_id:
                description: |
                  Name of the CI that sent the vote. If not set, no filter will
                  be set on CI.
                type: string
              sequential_mode:
                description: |
                  Use the sequential mode algorithm. In this case, return the
                  last tested repo within that timeframe for the CI job
                  described by previous_job_id. Defaults to false.
                type: boolean
              previous_job_id:
                description: |
                  If sequential_mode is set to true, look for jobs tested by
                  the CI identified by previous_job_id.
                type: string
          required: true
      # Expected responses for this operation:
      responses:
        # Response code
        201:
          description: Successful response
          # A schema describing your response object.
          # Use JSON Schema format
          schema:
            $ref: '#/definitions/Repo'
  /api/repo_status:
    # This is a HTTP operation
    get:
      # Describe this verb here. Note: you can use markdown
      description: |
        Get all the CI reports for a specific repository.
      parameters:
        - name: params
          in: body
          description: The JSON params to post
          required: true
          schema:
            type: object
            required:
              - commit_hash
              - distro_hash
            properties:
              commit_hash:
                description: |
                  commit_hash of the repo to fetch information for.
                type: string
              distro_hash:
                description: |
                  distro_hash of the repo to fetch information for.
                type: string
              success:
                description: |
                  If set to a value, only return the CI reports with the
                  specified vote. If not set, return all CI reports.
                type: boolean
      # Expected responses for this operation:
      responses:
        # Response code
        200:
          description: An array of votes
          schema:
            title: ArrayOfVotes
            type: array
            items:
              $ref: '#/definitions/CIVote'
  /api/civotes.html:
    get:
      description: |
        Get an overview of the CI reports.
      produces:
        - text/html
      parameters:
        - name: offset
          in: query
          description: Start after offset values. Defaults to 0.
          type: integer
      responses:
        # Response code
        200:
          description: An HTML page ith the CI votes.
  /api/civotes_detail.html:
    get:
      description: |
        Get a detailed overview of the CI reports for a specific repo, with optional filtering.
      produces:
        - text/html
      parameters:
        - name: commit_hash
          in: query
          description: |
            Filter on this commit hash.
            Either ci_name or a combination of commit_hash+distro_hash are required.
          type: string
        - name: distro_hash
          in: query
          description: |
            Filter on this distro hash.
            Either ci_name or a combination of commit_hash+distro_hash are required.
          type: string
        - name: ci_name
          in: query
          description: |
            Filter on this CI name.
            Either ci_name or a combination of commit_hash+distro_hash are required.
        - name: success
          in: query
          description: If set filter for successful/failed CI votes
          type: boolean
        - name: offset
          in: query
          description: Start after offset values. Defaults to 0.
          type: integer
      responses:
        # Response code
        200:
          description: An HTML page with the CI votes detail for a repo.
  /api/report.html:
    get:
      description: |
        Get a detailed report of the commit build status, with optional filtering.
      produces:
        - text/html
      parameters:
        - name: package
          in: query
          description: Filter on this package name.
          type: string
        - name: success
          in: query
          description: Filter on this successful/unsuccessful builds.
          type: boolean
        - name: offset
          in: query
          description: Start after offset values. Defaults to 0.
          type: integer
      responses:
        # Response code
        200:
          description: An HTML page with the report of the commit build status.
...
  /api/remote/import:
    post:
      security:
       - basicAuth: []
      description: |
        Import a commit built by another instance. This API call mimics the behavior of the
        ``dlrn-remote`` command, with the only exception of not being able to specify a custom
        rdoinfo location.
      parameters:
        - name: params
          in: body
          description: The JSON params to post
          required: true
          schema:
            $ref: '#/definitions/Import'
      responses:
        201:
          description: Successful response
          schema:
            $ref: '#/definitions/Import'
definitions:
...
  Repo:
    type: object
    properties:
      commit_hash:
        type: string
        description: commit_hash of tested repo
      distro_hash:
        type: string
        description: distro_hash of tested repo
      success:
        type: boolean
        description: whether the test was successful or not
      job_id:
        type: string
        description: name of the CI sending the vote
      in_progress:
        type: boolean
        description: is this CI job still in-progress?
      timestamp:
        type: integer
        description: Timestamp for this CI Vote (taken from the DLRN system time)
      user:
        type: string
        description: User who created this CI vote
...
  Import:
    type: object
    properties:
      repo_url:
        type: string
        description: Base repository URL for imported remote repo